Restructure occ api
