--- packages/occ-worker/src/occapi.ts
@@ -1,638 +1,0 @@
-import { OCC } from '@jupytercad/opencascade';
-import { v4 as uuid } from 'uuid';
-
-import {
-  IBox,
-  ICone,
-  ICut,
-  ICylinder,
-  IExtrusion,
-  IFuse,
-  IIntersection,
-  IJCadContent,
-  Parts,
-  ISketchObject,
-  ISphere,
-  ITorus,
-  IAny,
-  IPostOperator
-} from '@jupytercad/schema';
-
-import { getOcc } from './actions';
-import { _GeomCircle } from './geometry/geomCircle';
-import { _GeomLine } from './geometry/geomLineSegment';
-import { operatorCache } from './operatorcache';
-import { IAllOperatorFunc, IOperatorArg } from './types';
-import { toRad } from './utils';
-
-function setShapePlacement(
-  shape: OCC.TopoDS_Shape,
-  placement?: {
-    Position: number[];
-    Axis: number[];
-    Angle: number;
-  }
-): OCC.TopoDS_Shape {
-  if (!placement) {
-    return shape;
-  }
-  const oc = getOcc();
-  const trsf = new oc.gp_Trsf_1();
-
-  const ax = new oc.gp_Ax1_2(
-    new oc.gp_Pnt_3(0, 0, 0),
-    new oc.gp_Dir_4(placement.Axis[0], placement.Axis[1], placement.Axis[2])
-  );
-  const angle = toRad(placement.Angle);
-  trsf.SetRotation_1(ax, angle);
-  trsf.SetTranslationPart(
-    new oc.gp_Vec_4(
-      placement.Position[0],
-      placement.Position[1],
-      placement.Position[2]
-    )
-  );
-  const loc = new oc.TopLoc_Location_2(trsf);
-  shape.Location_2(loc, true);
-  return shape;
-}
-
-function _Box(arg: IBox, _: IJCadContent): OCC.TopoDS_Shape | undefined {
-  const { Length, Width, Height, Placement } = arg;
-  const oc = getOcc();
-  const box = new oc.BRepPrimAPI_MakeBox_2(Length, Width, Height);
-  const shape = box.Shape();
-  return setShapePlacement(shape, Placement);
-}
-
-function _Cylinder(
-  arg: ICylinder,
-  _: IJCadContent
-): OCC.TopoDS_Shape | undefined {
-  const { Radius, Height, Angle, Placement } = arg;
-  const oc = getOcc();
-  const cylinder = new oc.BRepPrimAPI_MakeCylinder_2(
-    Radius,
-    Height,
-    toRad(Angle)
-  );
-  const shape = cylinder.Shape();
-  return setShapePlacement(shape, Placement);
-}
-
-function _Sphere(arg: ISphere, _: IJCadContent): OCC.TopoDS_Shape | undefined {
-  const { Radius, Angle1, Angle2, Angle3, Placement } = arg;
-  const oc = getOcc();
-  const sphere = new oc.BRepPrimAPI_MakeSphere_4(
-    Radius,
-    toRad(Angle1),
-    toRad(Angle2),
-    toRad(Angle3)
-  );
-  const shape = sphere.Shape();
-  return setShapePlacement(shape, Placement);
-}
-
-function _Cone(arg: ICone, _: IJCadContent): OCC.TopoDS_Shape | undefined {
-  const { Radius1, Radius2, Height, Angle, Placement } = arg;
-  const oc = getOcc();
-  const cone = new oc.BRepPrimAPI_MakeCone_2(
-    Radius1,
-    Radius2,
-    Height,
-    toRad(Angle)
-  );
-  const shape = cone.Shape();
-  return setShapePlacement(shape, Placement);
-}
-
-function _Torus(arg: ITorus, _: IJCadContent): OCC.TopoDS_Shape | undefined {
-  const { Radius1, Radius2, Angle1, Angle2, Angle3, Placement } = arg;
-  const oc = getOcc();
-  const torus = new oc.BRepPrimAPI_MakeTorus_4(
-    Radius1,
-    Radius2,
-    toRad(Angle1),
-    toRad(Angle2),
-    toRad(Angle3)
-  );
-  const shape = torus.Shape();
-  return setShapePlacement(shape, Placement);
-}
-
-function _Cut(arg: ICut, content: IJCadContent): OCC.TopoDS_Shape | undefined {
-  const { Placement, Base, Tool } = arg;
-  const oc = getOcc();
-  const baseObject = content.objects.filter(obj => obj.name === Base);
-  const toolObject = content.objects.filter(obj => obj.name === Tool);
-  if (baseObject.length === 0 || toolObject.length === 0) {
-    return;
-  }
-  const baseShape = baseObject[0].shape;
-  const toolShape = toolObject[0].shape;
-  if (
-    baseShape &&
-    ShapesFactory[baseShape] &&
-    toolShape &&
-    ShapesFactory[toolShape]
-  ) {
-    const base = ShapesFactory[baseShape](
-      baseObject[0].parameters as IOperatorArg,
-      content
-    );
-    const tool = ShapesFactory[toolShape](
-      toolObject[0].parameters as IOperatorArg,
-      content
-    );
-    if (base && tool && base.occShape && tool.occShape) {
-      baseObject[0].visible = false;
-      toolObject[0].visible = false;
-      const operator = new oc.BRepAlgoAPI_Cut_3(
-        base.occShape,
-        tool.occShape,
-        new oc.Message_ProgressRange_1()
-      );
-      if (operator.IsDone()) {
-        return setShapePlacement(operator.Shape(), Placement);
-      }
-    }
-  }
-}
-
-function _Fuse(
-  arg: IFuse,
-  content: IJCadContent
-): OCC.TopoDS_Shape | undefined {
-  const oc = getOcc();
-  const { Shapes, Placement } = arg;
-  const occShapes: OCC.TopoDS_Shape[] = [];
-  Shapes.forEach(Base => {
-    const baseObject = content.objects.filter(obj => obj.name === Base);
-    if (baseObject.length === 0) {
-      return;
-    }
-    const baseShape = baseObject[0].shape;
-    if (baseShape && ShapesFactory[baseShape]) {
-      const base = ShapesFactory[baseShape](
-        baseObject[0].parameters as IOperatorArg,
-        content
-      );
-      if (base && base.occShape) {
-        occShapes.push(base.occShape);
-        baseObject[0].visible = false;
-      }
-    }
-  });
-  const operator = new oc.BRepAlgoAPI_Fuse_3(
-    occShapes[0],
-    occShapes[1],
-    new oc.Message_ProgressRange_1()
-  );
-  if (operator.IsDone()) {
-    return setShapePlacement(operator.Shape(), Placement);
-  }
-  return;
-}
-
-function _Intersection(
-  arg: IIntersection,
-  content: IJCadContent
-): OCC.TopoDS_Shape | undefined {
-  const oc = getOcc();
-  const { Shapes, Placement } = arg;
-  const occShapes: OCC.TopoDS_Shape[] = [];
-  Shapes.forEach(Base => {
-    const baseObject = content.objects.filter(obj => obj.name === Base);
-    if (baseObject.length === 0) {
-      return;
-    }
-    const baseShape = baseObject[0].shape;
-    if (baseShape && ShapesFactory[baseShape]) {
-      const base = ShapesFactory[baseShape](
-        baseObject[0].parameters as IOperatorArg,
-        content
-      );
-      if (base && base.occShape) {
-        occShapes.push(base.occShape);
-        baseObject[0].visible = false;
-      }
-    }
-  });
-  const operator = new oc.BRepAlgoAPI_Common_3(
-    occShapes[0],
-    occShapes[1],
-    new oc.Message_ProgressRange_1()
-  );
-  if (operator.IsDone()) {
-    return setShapePlacement(operator.Shape(), Placement);
-  }
-  return;
-}
-
-export function _SketchObject(
-  arg: ISketchObject,
-  content: IJCadContent
-): OCC.TopoDS_Shape | undefined {
-  const oc = getOcc();
-  const builder = new oc.BRep_Builder();
-  const compound = new oc.TopoDS_Compound();
-  if (arg.Geometry.length === 0) {
-    return undefined;
-  }
-  builder.MakeCompound(compound);
-  for (const geom of arg.Geometry) {
-    switch (geom.TypeId) {
-      case 'Part::GeomCircle':
-        builder.Add(compound, _GeomCircle(geom));
-        break;
-
-      case 'Part::GeomLineSegment': {
-        builder.Add(compound, _GeomLine(geom));
-        break;
-      }
-      default:
-        break;
-    }
-  }
-  return compound;
-}
-
-function _Extrude(
-  arg: IExtrusion,
-  content: IJCadContent
-): OCC.TopoDS_Shape | undefined {
-  const { Base, Dir, LengthFwd, LengthRev, Placement, Solid } = arg;
-  const oc = getOcc();
-  const baseObject = content.objects.filter(obj => obj.name === Base);
-
-  if (baseObject.length === 0) {
-    return;
-  }
-  const baseShape = baseObject[0].shape;
-
-  if (baseShape && ShapesFactory[baseShape]) {
-    const base = ShapesFactory[baseShape](
-      baseObject[0].parameters as IOperatorArg,
-      content
-    );
-    if (!base || !base.occShape) {
-      return;
-    }
-    const dirVec = new oc.gp_Vec_4(Dir[0], Dir[1], Dir[2]);
-    const vec = dirVec.Multiplied(LengthFwd + LengthRev);
-    let baseCopy = new oc.BRepBuilderAPI_Copy_2(
-      base.occShape,
-      true,
-      false
-    ).Shape();
-    if (LengthRev !== 0) {
-      const mov = new oc.gp_Trsf_1();
-      mov.SetTranslation_1(dirVec.Multiplied(-LengthRev));
-      const loc = new oc.TopLoc_Location_2(mov);
-      baseCopy.Move(loc, true);
-    }
-
-    if (Solid) {
-      const xp = new oc.TopExp_Explorer_2(
-        baseCopy,
-        oc.TopAbs_ShapeEnum.TopAbs_FACE as any,
-        oc.TopAbs_ShapeEnum.TopAbs_SHAPE as any
-      );
-      if (xp.More()) {
-        //source shape has faces. Just extrude as-is.
-      } else {
-        const wireEx = new oc.TopExp_Explorer_2(
-          baseCopy,
-          oc.TopAbs_ShapeEnum.TopAbs_EDGE as any,
-          oc.TopAbs_ShapeEnum.TopAbs_SHAPE as any
-        );
-        const wireMaker = new oc.BRepBuilderAPI_MakeWire_1();
-        while (wireEx.More()) {
-          const ed = oc.TopoDS.Edge_1(wireEx.Current());
-          wireMaker.Add_1(ed);
-          wireEx.Next();
-        }
-
-        const wire = wireMaker.Wire();
-        const faceMaker = new oc.BRepBuilderAPI_MakeFace_15(wire, false);
-        baseCopy = faceMaker.Face();
-      }
-    }
-    const result = new oc.BRepPrimAPI_MakePrism_1(baseCopy, vec, false, true);
-    return setShapePlacement(result.Shape(), Placement);
-  }
-
-  return;
-}
-
-export function _Any(
-  arg: IAny,
-  content: IJCadContent
-): OCC.TopoDS_Shape | undefined {
-  const { Content, Type, Placement } = arg;
-  const result = _loadObjectFile({ content: Content, type: Type });
-  if (result) {
-    return setShapePlacement(result, Placement);
-  }
-}
-
-export function _PostOperator(
-  arg: IPostOperator,
-  content: IJCadContent
-): { occBrep: string } {
-  const baseObject = content.objects.filter(obj => obj.name === arg.Object);
-  if (baseObject.length === 0) {
-    return { occBrep: '' };
-  }
-  const baseShape = baseObject[0].shape;
-  if (baseShape && ShapesFactory[baseShape]) {
-    const base = ShapesFactory[baseShape](
-      baseObject[0].parameters as IOperatorArg,
-      content
-    );
-    if (base?.occShape) {
-      const occBrep = _writeBrep(base.occShape);
-      return { occBrep };
-    }
-  }
-  return { occBrep: '' };
-}
-
-export function _loadBrepFile(content: string): OCC.TopoDS_Shape | undefined {
-  const oc = getOcc();
-  const fakeFileName = `${uuid()}.brep`;
-  oc.FS.createDataFile('/', fakeFileName, content, true, true, true);
-  const shape = new oc.TopoDS_Shape();
-  const builder = new oc.BRep_Builder();
-  const progress = new oc.Message_ProgressRange_1();
-  oc.BRepTools.Read_2(shape, fakeFileName, builder, progress);
-  oc.FS.unlink('/' + fakeFileName);
-  return shape;
-}
-
-export function _loadStepFile(content: string): OCC.TopoDS_Shape | undefined {
-  const oc = getOcc();
-  const fakeFileName = `${uuid()}.STEP`;
-  oc.FS.createDataFile('/', fakeFileName, content, true, true, true);
-
-  const reader = new oc.STEPControl_Reader_1();
-
-  const readResult = reader.ReadFile(fakeFileName);
-
-  if (readResult === oc.IFSelect_ReturnStatus.IFSelect_RetDone) {
-    reader.TransferRoots(new oc.Message_ProgressRange_1());
-    const shape = reader.OneShape();
-    oc.FS.unlink('/' + fakeFileName);
-    return shape;
-  } else {
-    console.error('Something in OCCT went wrong trying to read');
-    return undefined;
-  }
-}
-
-export function _loadObjectFile(arg: {
-  content: string;
-  type: IAny['Type'];
-}): OCC.TopoDS_Shape | undefined {
-  switch (arg.type.toLowerCase()) {
-    case 'brep':
-      return _loadBrepFile(arg.content);
-    case 'step':
-      return _loadStepFile(arg.content);
-    default:
-      throw `${arg.type} file not supported`;
-  }
-}
-
-export function _writeBrep(shape: OCC.TopoDS_Shape): string {
-  const oc = getOcc();
-  const fakeFileName = `${uuid()}.brep`;
-  const progress = new oc.Message_ProgressRange_1();
-  oc.BRepTools.Write_4(
-    shape,
-    fakeFileName,
-    false,
-    false,
-    oc.TopTools_FormatVersion.TopTools_FormatVersion_VERSION_1 as any,
-    progress
-  );
-  const value = oc.FS.readFile('/' + fakeFileName, { encoding: 'utf8' });
-  oc.FS.unlink('/' + fakeFileName);
-  return value;
-}
-
-export function makeShapeOnMesh(
-  myMeshHandle: OCC.Handle_Poly_Triangulation
-): OCC.TopoDS_Shape | undefined {
-  if (myMeshHandle.IsNull()) {
-    return;
-  }
-  const myMesh = myMeshHandle.get();
-  if (myMesh.NbNodes() === 0 || myMesh.NbTriangles() === 0) {
-    return;
-  }
-  const SMALL = 1e-9;
-  const oc = getOcc();
-  const aNbNodes = myMesh.NbNodes();
-  const aNbTriangles = myMesh.NbTriangles();
-
-  const aPnt2VertexMap: Map<number, OCC.TopoDS_Vertex> = new Map();
-  for (let idx = 1; idx <= aNbNodes; idx++) {
-    const aP = myMesh.Node(idx);
-    const aV = new OCC.BRepBuilderAPI_MakeVertex(aP);
-    aPnt2VertexMap.set(idx, aV.Vertex());
-  }
-
-  const anEdgeToTEgeMap: Map<[number, number], OCC.TopoDS_Edge> = new Map();
-
-  for (let idx = 1; idx <= aNbTriangles; idx++) {
-    const aTriangle = myMesh.Triangle(idx);
-    const anIdx: [number, number, number] = [
-      aTriangle.Value(1),
-      aTriangle.Value(2),
-      aTriangle.Value(3)
-    ];
-
-    if (
-      anIdx[0] === anIdx[1] ||
-      anIdx[0] === anIdx[2] ||
-      anIdx[1] === anIdx[2]
-    ) {
-      continue;
-    }
-    const aP1 = myMesh.Node(anIdx[0]);
-    const aP2 = myMesh.Node(anIdx[1]);
-    const aP3 = myMesh.Node(anIdx[2]);
-    const aD1 = aP1.SquareDistance(aP2);
-    const aD2 = aP1.SquareDistance(aP3);
-    const aD3 = aP2.SquareDistance(aP3);
-    if (aD1 < SMALL || aD2 < SMALL || aD3 < SMALL) {
-      continue;
-    }
-    const aV1 = aPnt2VertexMap.get(anIdx[0]);
-    const aV2 = aPnt2VertexMap.get(anIdx[1]);
-    const aV3 = aPnt2VertexMap.get(anIdx[2]);
-    if (!aV1 || !aV2 || !aV3) {
-      return;
-    }
-    const aMeshEdge1: [number, number] = [
-      Math.min(anIdx[0], anIdx[1]),
-      Math.max(anIdx[0], anIdx[1])
-    ];
-    const aMeshEdge2: [number, number] = [
-      Math.min(anIdx[1], anIdx[2]),
-      Math.max(anIdx[1], anIdx[2])
-    ];
-    const aMeshEdge3: [number, number] = [
-      Math.min(anIdx[2], anIdx[0]),
-      Math.max(anIdx[2], anIdx[0])
-    ];
-
-    const aMaker1 = new oc.BRepBuilderAPI_MakeEdge_2(aV1, aV2);
-    const aTE1 = aMaker1.Edge();
-    if (anIdx[1] < anIdx[0]) {
-      aTE1.Reverse();
-    }
-    const aMaker2 = new oc.BRepBuilderAPI_MakeEdge_2(aV2, aV3);
-    const aTE2 = aMaker2.Edge();
-    if (anIdx[2] < anIdx[1]) {
-      aTE2.Reverse();
-    }
-
-    const aMaker3 = new oc.BRepBuilderAPI_MakeEdge_2(aV3, aV1);
-    const aTE3 = aMaker3.Edge();
-    if (anIdx[0] < anIdx[2]) {
-      aTE3.Reverse();
-    }
-    anEdgeToTEgeMap.set(aMeshEdge1, aTE1);
-    anEdgeToTEgeMap.set(aMeshEdge2, aTE2);
-    anEdgeToTEgeMap.set(aMeshEdge3, aTE3);
-  }
-  const aResult = new oc.TopoDS_Compound();
-  const aBB = new oc.BRep_Builder();
-  aBB.MakeCompound(aResult);
-
-  for (let idx = 1; idx < aNbTriangles; idx++) {
-    const aTriangle = myMesh.Triangle(idx);
-    const anIdx: [number, number, number] = [
-      aTriangle.Value(1),
-      aTriangle.Value(2),
-      aTriangle.Value(3)
-    ];
-
-    const aMeshEdge1: [number, number] = [
-      Math.min(anIdx[0], anIdx[1]),
-      Math.max(anIdx[0], anIdx[1])
-    ];
-    const aMeshEdge2: [number, number] = [
-      Math.min(anIdx[1], anIdx[2]),
-      Math.max(anIdx[1], anIdx[2])
-    ];
-    const aMeshEdge3: [number, number] = [
-      Math.min(anIdx[2], anIdx[0]),
-      Math.max(anIdx[2], anIdx[0])
-    ];
-    const isReversed1 = anIdx[1] < anIdx[0];
-    const isReversed2 = anIdx[2] < anIdx[1];
-    const isReversed3 = anIdx[0] < anIdx[2];
-
-    const aHasAllEdges =
-      anEdgeToTEgeMap.has(aMeshEdge1) &&
-      anEdgeToTEgeMap.has(aMeshEdge2) &&
-      anEdgeToTEgeMap.has(aMeshEdge3);
-    if (!aHasAllEdges) {
-      continue;
-    }
-    const aTEdge1 = anEdgeToTEgeMap.get(aMeshEdge1);
-    const aTEdge2 = anEdgeToTEgeMap.get(aMeshEdge2);
-    const aTEdge3 = anEdgeToTEgeMap.get(aMeshEdge3);
-    if (!aTEdge1 || !aTEdge2 || !aTEdge3) {
-      continue;
-    }
-    if (isReversed1) {
-      aTEdge1.Reverse();
-    }
-    if (isReversed2) {
-      aTEdge2.Reverse();
-    }
-    if (isReversed3) {
-      aTEdge3.Reverse();
-    }
-    const aWireMaker = new oc.BRepBuilderAPI_MakeWire_1();
-    aWireMaker.Add_1(aTEdge1);
-    aWireMaker.Add_1(aTEdge2);
-    aWireMaker.Add_1(aTEdge3);
-    const aWire = aWireMaker.Wire();
-
-    const aC1 = new oc.BRepAdaptor_Curve_2(aTEdge1);
-    const aC2 = new oc.BRepAdaptor_Curve_2(aTEdge2);
-    const aD1 = aC1.Line().Direction();
-    const aD2 = aC2.Line().Direction();
-    const aN = aD1.XYZ().Crossed(aD2.XYZ());
-
-    if (aN.SquareModulus() < SMALL) {
-      continue;
-    }
-    if (aTEdge1.Orientation_1() === oc.TopAbs_Orientation.TopAbs_REVERSED) {
-      aN.Reverse();
-    }
-    if (aTEdge2.Orientation_1() === oc.TopAbs_Orientation.TopAbs_REVERSED) {
-      aN.Reverse();
-    }
-    const aNorm = new oc.gp_Dir_3(aN);
-    const aPln = new oc.gp_Pln_3(myMesh.Node(anIdx[0]), aNorm);
-    const aFaceMaker = new oc.BRepBuilderAPI_MakeFace_16(aPln, aWire, true);
-    const aFace = aFaceMaker.Face();
-    aBB.Add(aResult, aFace);
-  }
-  return aResult;
-}
-
-const Any = operatorCache<IAny>('Part::Any', _Any);
-const Box = operatorCache<IBox>('Part::Box', _Box);
-
-const Cylinder = operatorCache<ICylinder>('Part::Cylinder', _Cylinder);
-
-const Sphere = operatorCache<ISphere>('Part::Sphere', _Sphere);
-
-const Cone = operatorCache<ICone>('Part::Cone', _Cone);
-
-const Torus = operatorCache<ITorus>('Part::Torus', _Torus);
-
-const SketchObject = operatorCache<ISketchObject>(
-  'Sketcher::SketchObject',
-  _SketchObject
-);
-
-const Cut = operatorCache<ICut>('Part::Cut', _Cut);
-
-const Fuse = operatorCache<IFuse>('Part::MultiFuse', _Fuse);
-
-const Intersection = operatorCache<IIntersection>(
-  'Part::MultiCommon',
-  _Intersection
-);
-
-const Extrude = operatorCache<IExtrusion>('Part::Extrusion', _Extrude);
-
-export const ObjectFile = operatorCache<{
-  content: string;
-  type: IAny['Type'];
-}>('ObjectFile', _loadObjectFile);
-
-export const ShapesFactory: {
-  [key in Parts]: IAllOperatorFunc;
-} = {
-  'Part::Any': Any,
-  'Part::Box': Box,
-  'Part::Cylinder': Cylinder,
-  'Part::Sphere': Sphere,
-  'Part::Cone': Cone,
-  'Part::Torus': Torus,
-  'Part::Cut': Cut,
-  'Part::MultiFuse': Fuse,
-  'Part::Extrusion': Extrude,
-  'Part::MultiCommon': Intersection,
-  'Sketcher::SketchObject': SketchObject,
-  'Post::Operator': _PostOperator
-};
